requirements.txt
langchain
langchain-core
langchain-ollama
langchain-chroma
chromadb
pymysql
requests
python-dotenv